make all rem values be rem strings
