+++ libs/@guardian/source/src/foundations/utils/convert-value.ts
@@ -5,12 +5,3 @@ export const rootPixelFontSize = 16;
 
 export const pxToRem = (px: number): RemString =>
 	`${px / rootPixelFontSize}rem`;
-
-export const pxStringToRem = (px: PxString): RemString =>
-	pxToRem(pxStringToNumber(px));
-
-export const pxStringToNumber = <N extends number>(px: PxString<N>): N =>
-	Number(px.slice(0, -2)) as N;
-
-export const fontArrayToString = (fonts: readonly string[]): string =>
-	fonts.map((name) => (name.includes(' ') ? `"${name}"` : name)).join(', ');
